+++ bin/commands/BlastCommand.js
@@ -1,0 +1,7 @@
+const BlastCommand = {
+    blast: async function(interaction, channel) {
+        await interaction.reply('Hey from the blast command.');
+    }
+};
+
+module.exports = BlastCommand;